requirements.txt
requests==2.0
boto3==1.1.1
paramiko==1.15.2
scp==0.10.2
pyyaml==3.11
